extract shared daemon HTTP client, deduplicate JWT minting
